+++ app/src/main/java/activity/SearchActivity.kt
@@ -24,6 +24,7 @@ import api.dtos.TrackDto
 import api.dtos.TracksResponseDto
 import api.service.ITunesClient
 import com.example.playlistmaker.R
+import presentation.ui.player.PlayerActivity
 import storage.HISTORY_SHARED_PREFS
 import storage.HistoryStorage
 import retrofit2.Call
